@@ -1,172 +1,171 @@
-﻿namespace PgBrewer
+﻿namespace PgBrewer;
+
+using System.Collections.Generic;
+using System.ComponentModel;
+using System.IO;
+using System.Runtime.CompilerServices;
+
+public class Alcohol : INotifyPropertyChanged
 {
-    using System.Collections.Generic;
-    using System.ComponentModel;
-    using System.IO;
-    using System.Runtime.CompilerServices;
+    #region Init
+    public static Alcohol None { get; } = new Alcohol(string.Empty);
 
-    public class Alcohol : INotifyPropertyChanged
+    public Alcohol(string name)
     {
-        #region Init
-        public static Alcohol None { get; } = new Alcohol(string.Empty);
+        Name = name;
+        EffectList = DataArchive.ReadEffectList(name);
+        Previous = None;
+        Next = None;
+    }
+    #endregion
 
-        public Alcohol(string name)
+    #region Properties
+    public string Name { get; }
+    public List<Effect> EffectList { get; } = new List<Effect>();
+    public AlcoholLineCollection Lines { get; } = new AlcoholLineCollection();
+    public Alcohol Previous { get; private set; }
+    public Alcohol Next { get; private set; }
+    public List<ComponentAssociationCollection> PreviousToNext { get; private set; } = new List<ComponentAssociationCollection>();
+
+    public int MismatchCount
+    {
+        get => MismatchCountInternal;
+        set
         {
-            Name = name;
-            EffectList = DataArchive.ReadEffectList(name);
-            Previous = None;
-            Next = None;
-        }
-        #endregion
-
-        #region Properties
-        public string Name { get; }
-        public List<Effect> EffectList { get; } = new List<Effect>();
-        public AlcoholLineCollection Lines { get; } = new AlcoholLineCollection();
-        public Alcohol Previous { get; private set; }
-        public Alcohol Next { get; private set; }
-        public List<ComponentAssociationCollection> PreviousToNext { get; private set; } = new List<ComponentAssociationCollection>();
-
-        public int MismatchCount
-        {
-            get => MismatchCountInternal;
-            set
+            if (MismatchCountInternal != value)
             {
-                if (MismatchCountInternal != value)
-                {
-                    MismatchCountInternal = value;
-                    NotifyThisPropertyChanged();
-                }
+                MismatchCountInternal = value;
+                NotifyThisPropertyChanged();
             }
         }
+    }
 
-        private int MismatchCountInternal;
-        #endregion
+    private int MismatchCountInternal;
+    #endregion
 
-        #region Client Interface
-        public virtual void Save()
+    #region Client Interface
+    public virtual void Save()
+    {
+        List<int> Indexes = new List<int>();
+
+        for (int i = 0; i < Lines.Count; i++)
+            Indexes.Add(Lines[i].EffectIndex);
+
+        DataArchive.SetIndexList(Name, Indexes);
+    }
+
+    public virtual void ClearCalculateIndexes()
+    {
+        foreach (AlcoholLine Line in Lines)
+            Line.ClearCalculateIndex();
+    }
+
+    public virtual void RecalculateMismatchCount()
+    {
+        int NewCount = 0;
+
+        foreach (AlcoholLine Line in Lines)
+            if (!Line.IsMatching)
+                NewCount++;
+
+        MismatchCount = NewCount;
+    }
+
+    public virtual void Export(StreamWriter writer)
+    {
+        writer.WriteLine(Name);
+
+        foreach (AlcoholLine Line in Lines)
         {
-            List<int> Indexes = new List<int>();
+            string ExportedComponents = Line.GetExportedComponents();
+            int EffectIndex = Line.EffectIndex;
 
-            for (int i = 0; i < Lines.Count; i++)
-                Indexes.Add(Lines[i].EffectIndex);
-
-            DataArchive.SetIndexList(Name, Indexes);
+            if (EffectIndex >= 0)
+                writer.WriteLine($"{ExportedComponents};{EffectList[EffectIndex].Text}");
+            else
+                writer.WriteLine($"{ExportedComponents};");
         }
 
-        public virtual void ClearCalculateIndexes()
+        writer.WriteLine();
+    }
+
+    public virtual bool Import(StreamReader reader, ref int changeCount)
+    {
+        if (Name != reader.ReadLine())
+            return false;
+
+        for (int i = 0; i < Lines.Count; i++)
         {
-            foreach (AlcoholLine Line in Lines)
-                Line.ClearCalculateIndex();
-        }
+            AlcoholLine Line = Lines[i];
+            string LineString = reader.ReadLine()!;
 
-        public virtual void RecalculateMismatchCount()
-        {
-            int NewCount = 0;
+            string ExportedComponents = Line.GetExportedComponents();
+            ExportedComponents += ";";
 
-            foreach (AlcoholLine Line in Lines)
-                if (!Line.IsMatching)
-                    NewCount++;
-
-            MismatchCount = NewCount;
-        }
-
-        public virtual void Export(StreamWriter writer)
-        {
-            writer.WriteLine(Name);
-
-            foreach (AlcoholLine Line in Lines)
-            {
-                string ExportedComponents = Line.GetExportedComponents();
-                int EffectIndex = Line.EffectIndex;
-
-                if (EffectIndex >= 0)
-                    writer.WriteLine($"{ExportedComponents};{EffectList[EffectIndex].Text}");
-                else
-                    writer.WriteLine($"{ExportedComponents};");
-            }
-
-            writer.WriteLine();
-        }
-
-        public virtual bool Import(StreamReader reader, ref int changeCount)
-        {
-            if (Name != reader.ReadLine())
+            if (!LineString.StartsWith(ExportedComponents))
                 return false;
 
-            for (int i = 0; i < Lines.Count; i++)
+            LineString = LineString.Substring(ExportedComponents.Length);
+            if (LineString.Length > 0)
             {
-                AlcoholLine Line = Lines[i];
-                string LineString = reader.ReadLine()!;
-
-                string ExportedComponents = Line.GetExportedComponents();
-                ExportedComponents += ";";
-
-                if (!LineString.StartsWith(ExportedComponents))
-                    return false;
-
-                LineString = LineString.Substring(ExportedComponents.Length);
-                if (LineString.Length > 0)
+                int SelectedLineIndex = -1;
+                for (int j = 0; j < EffectList.Count; j++)
                 {
-                    int SelectedLineIndex = -1;
-                    for (int j = 0; j < EffectList.Count; j++)
+                    if (LineString == EffectList[j].Text)
                     {
-                        if (LineString == EffectList[j].Text)
-                        {
-                            SelectedLineIndex = j;
-                            break;
-                        }
-                    }
-
-                    if (SelectedLineIndex < 0)
-                        return false;
-
-                    if (Line.EffectIndex != SelectedLineIndex)
-                    {
-                        Line.EffectIndex = SelectedLineIndex;
-                        changeCount++;
+                        SelectedLineIndex = j;
+                        break;
                     }
                 }
-                else if (Line.EffectIndex >= 0)
+
+                if (SelectedLineIndex < 0)
+                    return false;
+
+                if (Line.EffectIndex != SelectedLineIndex)
                 {
-                    Line.EffectIndex = -1;
+                    Line.EffectIndex = SelectedLineIndex;
                     changeCount++;
                 }
             }
-
-            reader.ReadLine();
-
-            return true;
+            else if (Line.EffectIndex >= 0)
+            {
+                Line.EffectIndex = -1;
+                changeCount++;
+            }
         }
 
-        public static void Chain(Alcohol previous, Alcohol next, List<ComponentAssociationCollection> previousToNext)
-        {
-            previous.Next = next;
-            next.Previous = previous;
+        reader.ReadLine();
 
-            previous.PreviousToNext = previousToNext;
-        }
-        #endregion
-
-        #region Implementation of INotifyPropertyChanged
-        public event PropertyChangedEventHandler? PropertyChanged;
-
-        public void NotifyPropertyChanged(string propertyName)
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-
-        public void NotifyThisPropertyChanged([CallerMemberName] string propertyName = "")
-        {
-            PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
-        }
-        #endregion
-
-        #region Debugging
-        public override string ToString()
-        {
-            return Name;
-        }
-        #endregion
+        return true;
     }
+
+    public static void Chain(Alcohol previous, Alcohol next, List<ComponentAssociationCollection> previousToNext)
+    {
+        previous.Next = next;
+        next.Previous = previous;
+
+        previous.PreviousToNext = previousToNext;
+    }
+    #endregion
+
+    #region Implementation of INotifyPropertyChanged
+    public event PropertyChangedEventHandler? PropertyChanged;
+
+    public void NotifyPropertyChanged(string propertyName)
+    {
+        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+    }
+
+    public void NotifyThisPropertyChanged([CallerMemberName] string propertyName = "")
+    {
+        PropertyChanged?.Invoke(this, new PropertyChangedEventArgs(propertyName));
+    }
+    #endregion
+
+    #region Debugging
+    public override string ToString()
+    {
+        return Name;
+    }
+    #endregion
 }
