@@ -1,10 +1,9 @@
-﻿namespace PgBrewer
+﻿namespace PgBrewer;
+
+public partial class Alcoholx4x3x4x3Control : AlcoholControl
 {
-    public partial class Alcoholx4x3x4x3Control : AlcoholControl
+    public Alcoholx4x3x4x3Control()
     {
-        public Alcoholx4x3x4x3Control()
-        {
-            InitializeComponent();
-        }
+        InitializeComponent();
     }
 }
