@@ -1,10 +1,9 @@
-﻿namespace PgBrewer
+﻿namespace PgBrewer;
+
+public partial class Alcoholx4x4Control : AlcoholControl
 {
-    public partial class Alcoholx4x4Control : AlcoholControl
+    public Alcoholx4x4Control()
     {
-        public Alcoholx4x4Control()
-        {
-            InitializeComponent();
-        }
+        InitializeComponent();
     }
 }
