@@ -1,20 +1,19 @@
-﻿namespace PgBrewer
-{
-    public class Effect
-    {
-        #region Init
-        public Effect(string text, string? prefix, string? suffix)
-        {
-            Text = text;
-            Prefix = prefix;
-            Suffix = suffix;
-        }
-        #endregion
+﻿namespace PgBrewer;
 
-        #region Properties
-        public string Text { get; }
-        public string? Prefix { get; }
-        public string? Suffix { get; }
-        #endregion
+public class Effect
+{
+    #region Init
+    public Effect(string text, string? prefix, string? suffix)
+    {
+        Text = text;
+        Prefix = prefix;
+        Suffix = suffix;
     }
+    #endregion
+
+    #region Properties
+    public string Text { get; }
+    public string? Prefix { get; }
+    public string? Suffix { get; }
+    #endregion
 }
